cleaned up functions file
